+++ MapboxCoreNavigation/MBRouteController.m
@@ -12,6 +12,8 @@ const MBRouteControllerNotificationUserInfoKey MBRouteControllerRouteProgressKey
 const MBRouteControllerNotificationUserInfoKey MBRouteControllerLocationKey                   = @"location";
 const MBRouteControllerNotificationUserInfoKey MBRouteControllerRawLocationKey                = @"rawLocation";
 const MBRouteControllerNotificationUserInfoKey MBRouteControllerRoutingErrorKey               = @"error";
+const MBRouteControllerNotificationUserInfoKey MBRouteControllerVisualInstructionKey          = @"visualInstruction";
+const MBRouteControllerNotificationUserInfoKey MBRouteControllerSpokenInstructionKey          = @"spokenInstruction";
 const MBRouteControllerNotificationUserInfoKey MBRouteControllerIsProactiveKey                = @"RouteControllerDidFindFasterRoute";
 
 NSString *const MBErrorDomain = @"ErrorDomain";