+++ MapboxCoreNavigation/NavigationService.swift
@@ -458,6 +458,14 @@ extension MapboxNavigationService: RouterDelegate {
         delegate?.navigationService?(self, didUpdate: progress, with: location, rawLocation: rawLocation)
     }
     
+    public func router(_ router: Router, didPassVisualInstructionPoint instruction: VisualInstructionBanner, routeProgress: RouteProgress) {
+        delegate?.navigationService?(self, didPassVisualInstructionPoint: instruction, routeProgress: routeProgress)
+    }
+    
+    public func router(_ router: Router, didPassSpokenInstructionPoint instruction: SpokenInstruction, routeProgress: RouteProgress) {
+        delegate?.navigationService?(self, didPassSpokenInstructionPoint: instruction, routeProgress: routeProgress)
+    }
+    
     //MARK: Questions
     public func router(_ router: Router, shouldRerouteFrom location: CLLocation) -> Bool {
         return delegate?.navigationService?(self, shouldRerouteFrom: location) ?? Default.shouldRerouteFromLocation